Llaves foraneas delete on cascade, cambio datatype, arreglado insert em.

+++ src/main/java/cl/tbdlab/voluntariadoGrupo1/repositories/EmergenciaRepository.java
@@ -5,9 +5,9 @@ import java.util.List;
 
 public interface EmergenciaRepository {
     public int insertEmergencia(Emergencia emergencia);
-    public Emergencia readEmergencia(int id);
+    public Emergencia readEmergencia(Long id);
     public List<Emergencia> readEmergencia();
     public int updateEmergencia(Emergencia id);
-    public int deleteEmergencia(int id);
+    public int deleteEmergencia(Long id);
     public int deleteEmergencia();
 }
